Updated to 2014b

+++ timezone.spec
@@ -1,17 +1,18 @@
-#define tzdata_version %{version}
-%define tzdata_version 2013i
-%define tzcode_version 2013i
+%define tzdata_version 2014b
+%define tzcode_version 2014b
 
 # the zic(8) and zdump(8) manpages are already in man-pages
-%bcond_with manpages
-%ifarch %{ix86} x86_64 %arm
-%bcond_without java
+%define build_manpages 0
+%ifarch %mips %arm
+%define build_java 0
+%else
+%define build_java 1
 %endif
 
 Summary:	Timezone data
 Name:		timezone
-Epoch:		7
-Version:	2013i
+Epoch:		6
+Version:	2014b
 Release:	1
 License:	GPL
 Group:		System/Base
@@ -19,27 +20,24 @@ Source0:	ftp://ftp.iana.org/tz/releases/tzdata%{tzdata_version}.tar.gz
 Source1:	ftp://ftp.iana.org/tz/releases/tzcode%{tzcode_version}.tar.gz
 Source2:	javazic.tar.gz
 Source3:	update-localtime.sh
-Patch0:		tzdata-mdvconfig.patch
 Patch1:		tzdata-extra-tz-links.patch
 Patch2:		javazic-fixup.patch
-Patch3:		timezone-data-2013f-makefile.patch
-BuildRequires:	gawk
-BuildRequires:	perl
+Patch3:         javazic-exclusion-fix.patch
 Provides:	tzdata = %{version}-%{release}
+Conflicts:     %{name} < 6:2013f-1
+BuildRequires:	gawk, perl
 
 %description
 This package contains data files with rules for various timezones
 around the world.
 
-%if %{with java}
+%if %{build_java}
 %package java
 Summary:	Timezone data for Java
 Group:		System/Base
 Provides:	tzdata-java = %{version}-%{release}
-# We use gcj instead of OpenJDK to avoid a circular build dependency.
-# OpenJDK requires tzdata-java to be installed.
-#BuildRequires:	java-1.5.0-gcj-devel gcj-tools libgcj13
 BuildRequires:	java-rpmbuild
+BuildRequires:  java-devel
 
 %description java
 This package contains timezone information for use by Java runtimes.
@@ -49,12 +47,12 @@ This package contains timezone information for use by Java runtimes.
 %setup -q -c -a 1
 %patch1 -p1 -b .extra-tz-links
 
-%if %{with java}
+%if %{build_java}
 mkdir javazic
 tar xf %{SOURCE2} -C javazic
 pushd javazic
 %patch2 -p0 -b .javazic-fixup
-
+%patch3
 # Hack alert! sun.tools may be defined and installed in the
 # VM. In order to guarantee that we are using IcedTea/OpenJDK
 # for creating the zoneinfo files, rebase all the packages
@@ -88,41 +86,44 @@ popd
 	fi
 	rm -f zone.tab.new
 %endif
-%patch3 -p1
 
 %build
-%make TZDIR=%{_datadir}/zoneinfo CFLAGS="%{optflags} -std=gnu99" LDFLAGS="%{ldflags}"
+
+%make TZDIR=%{_datadir}/zoneinfo CFLAGS="%{optflags} -std=gnu99"
+
 grep -v tz-art.htm tz-link.htm > tz-link.html
 
-%if %{with java}
+%if %{build_java}
 pushd javazic
 %{javac} -source 1.5 -target 1.5 -classpath . `find . -name \*.java`
 popd
 %{java} -classpath javazic/ rht.tools.javazic.Main -V %{version} \
- -d zoneinfo/java \
- africa antarctica asia australasia europe northamerica pacificnew \
- southamerica backward etcetera solar87 solar88 solar89 systemv \
- javazic/tzdata_jdk/gmt javazic/tzdata_jdk/jdk11_backward
+  -d zoneinfo/java \
+  africa antarctica asia australasia europe northamerica pacificnew \
+  southamerica backward etcetera systemv \
+  javazic/tzdata_jdk/gmt javazic/tzdata_jdk/jdk11_backward
 %endif
 
 %install
-make TOPDIR="%{buildroot}/usr" \
-    TZDIR=%{buildroot}%{_datadir}/zoneinfo \
-    ETCDIR=%{buildroot}%{_sbindir} \
-    install
-
-%if %{with java}
-cp -a zoneinfo/java %{buildroot}%{_datadir}/javazi
-%endif
+make TOPDIR=%{buildroot}/usr \
+     TZDIR=%{buildroot}%{_datadir}/zoneinfo \
+     ETCDIR=%{buildroot}%{_sbindir} \
+     install
+rm -f %{buildroot}%{_datadir}/zoneinfo-posix
+ln -s . %{buildroot}%{_datadir}/zoneinfo/posix
+mv %{buildroot}%{_datadir}/zoneinfo-leaps %{buildroot}%{_datadir}/zoneinfo/right
 
 # nuke unpackaged files
 rm -f %{buildroot}%{_datadir}/zoneinfo/localtime
+rm -f %{buildroot}%{_sbindir}/tzselect
 rm -rf %{buildroot}/usr/{lib,man}
-rm -f %{buildroot}%{_bindir}/tzselect
-rm -f %{buildroot}%{_mandir}/man3/new*.3*
+
+%if %{build_java}
+cp -a zoneinfo/java %{buildroot}%{_datadir}/javazi
+%endif
 
 # install man pages
-%if %{with manpages}
+%if %{build_manpages}
 mkdir -p %{buildroot}%{_mandir}/man8
 for f in zic zdump; do
 install -m 644 $f.8 %{buildroot}%{_mandir}/man8/
@@ -136,13 +137,12 @@ perl -pi -e 's|\@datadir\@|%{_datadir}|;' \
 	 -e 's|\@sysconfdir\@|%{_sysconfdir}|' \
 	%{buildroot}%{_sbindir}/update-localtime
 
+%post -p %{_sbindir}/update-localtime
+
 %pretrans
-# This is being replaced with a symlink to .
-if [ ! -L %{_datadir}/zoneinfo/posix ]; then
+if [ -e %{_datadir}/zoneinfo/posix -a ! -L %{_datadir}/zoneinfo/posix ]; then
   rm -rf %{_datadir}/zoneinfo/posix
 fi
-
-%posttrans -p %{_sbindir}/update-localtime
 
 %files
 %doc README
@@ -151,14 +151,14 @@ fi
 %{_sbindir}/zdump
 %{_sbindir}/zic
 %{_sbindir}/update-localtime
-%if %{with manpages}
+%if %{build_manpages}
 %{_mandir}/man8/zdump.8*
 %{_mandir}/man8/zic.8*
 %endif
 %dir %{_datadir}/zoneinfo
 %{_datadir}/zoneinfo/*
 
-%if %{with java}
+%if %{build_java}
 %files java
 %{_datadir}/javazi
 %endif
